Extract RcWriter to shared code.
