+++ ailets-rs/actor_runtime_mocked/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod rc_writer;
 pub mod vfs;
 
+pub use rc_writer::RcWriter;
 pub use vfs::{add_file, clear_mocks, get_file, IO_INTERRUPT, WANT_ERROR};
